Fixes #13782: Fix tag exclusion support for contact assignments
